requirements.txt
numpy==2.1.3
pyserial==3.5
scipy==1.14.1
